src/apps/places/http/serializers.py
from rest_framework import serializers


class PlaceSerializer(serializers.Serializer):
    name = serializers.CharField()
    latitude = serializers.FloatField()
    longitude = serializers.FloatField()
