requirements.txt
asgiref==3.11.1
attrs==25.4.0
black==26.1.0
certifi==2026.1.4
click==8.3.1
Django==5.2.11
django-environ==0.12.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
geopandas==1.1.2
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
librt==0.8.0
mypy==1.19.1
mypy_extensions==1.1.0
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pathspec==1.0.4
platformdirs==4.9.2
psycopg2-binary==2.9.11
PyJWT==2.11.0
pyogrio==0.12.1
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
ruff==0.15.1
shapely==2.1.2
six==1.17.0
sqlparse==0.5.5
tomli==2.4.0
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
